Extract PS_nested related funs into a different file.
Revert some arg-to-kwarg changes
