+++ src/utils.jl
@@ -496,12 +496,12 @@ end
 
 # --------------------------------------------------------------------------------------------------
 """
-    R = rescale(A, a=0.0, b=1.0; inputmin=nothing, inputmax=nothing, stretch=false, type=nothing)
+    R = rescale(A; low=0.0, up=1.0; inputmin=nothing, inputmax=nothing, stretch=false, type=nothing)
 
 - `A`: is either a GMTgrid, GMTimage, Matrix{AbstractArray} or a file name. In later case the file is read
    with a call to `gmtread` that automatically decides how to read it based on the file extension ... not 100% safe.
 - `rescale(A)` rescales all entries of an array `A` to [0,1].
-- `rescale(A,b,c)` rescales all entries of A to the interval [b,c].
+- `rescale(A,low,up)` rescales all entries of A to the interval [low,up].
 - `rescale(..., inputmin=imin)` sets the lower bound `imin` for the input range. Input values less
    than `imin` will be replaced with `imin`. The default is min(A).
 - `rescale(..., inputmax=imax)` sets the lower bound `imax` for the input range. Input values greater
@@ -510,7 +510,7 @@ end
    will (try to) find good limits for histogram stretching. The form `stretch=(imin,imax)` allows specifying
    the input limits directly.
 - `type`: Converts the scaled array to this data type. Valid options are all Unsigned types (e.g. `UInt8`).
-   Default returns the same data type as `A` if it's an AbstractFloat, or Flot64 if `A` is an integer.
+   Default returns the same data type as `A` if it's an AbstractFloat, or Float64 if `A` is an integer.
 
 Returns a GMTgrid if `A` is a GMTgrid of floats, a GMTimage if `A` is a GMTimage and `type` is used or
 an array of Float32|64 otherwise.
@@ -1252,3 +1252,5 @@ end
 =#
 
 #GI.geometry[1].geoms[1].rings[1].vertices.data[1].coords.lat.val
+
+#include("imreconstruct.jl")
